Fixed submatch results in case of empty last repetition.
Added a boolean flag to each submatch group. The flag indicates in this
group is "frozen" (i.e. its value is final and will never change). If
frozen flag is not set and the algorithm finds a nonempty submatch, then
the previous submatch should be overwritten with the new one.

"Frozen" flag is set in the following way. When we find a nonempty match
for some group, we loop over all nested groups and freeze them. This
means that we have to keep a range of nested submatch indices for each
group. (A range, that is, two numbers, is enough because groups are
enumerated contiguously).